# src/config/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Smart University Communication and Venue Notification System API",
    "version": "1.0.0",
    "description": "A comprehensive backend system for university communication, announcements, venue management, and real-time notifications",
    "contact": {
      "name": "University of Zimbabwe",
      "email": "support@uz.ac.zw"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
...
  "paths": {
...
    "/api/users": {
      "get": {
        "tags": ["Users"],
        "summary": "Get all users (Admin only)",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"name": "role", "in": "query", "schema": {"type": "string"}},
          {"name": "department", "in": "query", "schema": {"type": "string"}},
          {"name": "search", "in": "query", "schema": {"type": "string"}},
          {"name": "page", "in": "query", "schema": {"type": "integer", "default": 1}},
          {"name": "limit", "in": "query", "schema": {"type": "integer", "default": 10}}
        ],
        "responses": {
          "200": {"description": "Users retrieved successfully"},
          "403": {"description": "Forbidden"}
        }
      }
    },
    "/api/users/{id}": {
      "get": {
        "tags": ["Users"],
        "summary": "Get user by ID",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"name": "id", "in": "path", "required": true, "schema": {"type": "string"}}
        ],
        "responses": {
          "200": {"description": "User retrieved"},
          "404": {"description": "User not found"}
        }
      },
      "put": {
        "tags": ["Users"],
        "summary": "Update user",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"name": "id", "in": "path", "required": true, "schema": {"type": "string"}}
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "firstName": {"type": "string"},
                  "lastName": {"type": "string"},
                  "department": {"type": "string"},
                  "password": {"type": "string"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {"description": "User updated"},
          "403": {"description": "Forbidden"}
        }
      },
      "delete": {
        "tags": ["Users"],
        "summary": "Delete user (Admin only)",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"name": "id", "in": "path", "required": true, "schema": {"type": "string"}}
        ],
        "responses": {
          "200": {"description": "User deleted"},
          "403": {"description": "Forbidden"}
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}